@@ -1,6 +1,6 @@
 {
   "name": "@lenne.tech/nest-server",
-  "version": "9.1.0",
+  "version": "9.2.0",
   "description": "Modern, fast, powerful Node.js web framework in TypeScript based on Nest with a GraphQL API and a connection to MongoDB (or other databases).",
   "keywords": [
     "node",
@@ -60,78 +60,83 @@
     "node": ">= 16.13.0"
   },
   "dependencies": {
-    "@apollo/gateway": "2.2.3",
-    "@nestjs/apollo": "10.1.7",
-    "@nestjs/common": "9.2.1",
-    "@nestjs/core": "9.2.1",
-    "@nestjs/graphql": "10.1.7",
-    "@nestjs/jwt": "10.0.1",
+    "@apollo/gateway": "2.3.2",
+    "@nestjs/apollo": "10.2.0",
+    "@nestjs/common": "9.3.9",
+    "@nestjs/core": "9.3.9",
+    "@nestjs/graphql": "10.2.0",
+    "@nestjs/jwt": "10.0.2",
     "@nestjs/mongoose": "9.2.1",
-    "@nestjs/passport": "9.0.0",
-    "@nestjs/platform-express": "9.2.1",
-    "@nestjs/schedule": "2.1.0",
+    "@nestjs/passport": "9.0.3",
+    "@nestjs/platform-express": "9.3.9",
+    "@nestjs/schedule": "2.2.0",
     "apollo-server-core": "3.11.1",
     "apollo-server-express": "3.11.1",
     "bcrypt": "5.1.0",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.0",
+    "compression": "1.7.4",
+    "cookie-parser": "1.4.6",
     "ejs": "3.1.8",
     "graphql": "16.6.0",
+    "graphql-query-complexity": "0.12.0",
     "graphql-subscriptions": "2.0.0",
     "graphql-upload": "15.0.2",
     "js-sha256": "0.9.0",
-    "json-to-graphql-query": "2.2.4",
+    "json-to-graphql-query": "2.2.5",
     "light-my-request": "5.8.0",
     "lodash": "4.17.21",
-    "mongodb": "4.13.0",
-    "mongoose": "6.8.4",
+    "mongodb": "4.14.0",
+    "mongoose": "6.9.1",
     "mongoose-gridfs": "1.3.0",
-    "multer-gridfs-storage": "5.0.2",
     "multer": "1.4.5-lts.1",
-    "node-mailjet": "6.0.1",
-    "nodemailer": "6.9.0",
+    "multer-gridfs-storage": "5.0.2",
+    "node-mailjet": "6.0.2",
+    "nodemailer": "6.9.1",
     "nodemon": "2.0.20",
     "passport": "0.6.0",
     "passport-jwt": "4.0.1",
     "reflect-metadata": "0.1.13",
     "rfdc": "1.3.0",
-    "rimraf": "4.1.1",
+    "rimraf": "4.1.2",
     "rxjs": "7.8.0"
   },
   "devDependencies": {
-    "@nestjs/testing": "9.2.1",
+    "@nestjs/testing": "9.3.9",
+    "@types/compression": "1.7.2",
+    "@types/cookie-parser": "1.4.3",
     "@types/cron": "2.0.0",
     "@types/ejs": "3.1.1",
-    "@types/jest": "29.2.6",
+    "@types/jest": "29.4.0",
     "@types/lodash": "4.14.191",
     "@types/multer": "1.4.7",
-    "@types/node": "18.11.18",
+    "@types/node": "18.13.0",
     "@types/nodemailer": "6.4.7",
     "@types/passport": "1.0.11",
     "@types/supertest": "2.0.12",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
+    "@typescript-eslint/eslint-plugin": "5.52.0",
+    "@typescript-eslint/parser": "5.52.0",
     "coffeescript": "2.7.0",
-    "eslint": "8.32.0",
+    "eslint": "8.34.0",
     "eslint-config-prettier": "8.6.0",
     "find-file-up": "2.0.1",
-    "grunt": "1.5.3",
+    "grunt": "1.6.1",
     "grunt-bg-shell": "2.3.3",
     "grunt-contrib-clean": "2.0.1",
     "grunt-contrib-watch": "1.1.0",
     "grunt-sync": "0.8.2",
     "husky": "8.0.3",
-    "jest": "29.3.1",
+    "jest": "29.4.2",
     "npm-watch": "0.11.0",
     "pm2": "5.2.2",
-    "prettier": "2.8.3",
+    "prettier": "2.8.4",
     "pretty-quick": "3.1.3",
     "supertest": "6.3.3",
     "ts-jest": "29.0.5",
     "ts-morph": "17.0.1",
     "ts-node": "10.9.1",
     "tsconfig-paths": "4.1.2",
-    "typescript": "4.9.4",
+    "typescript": "4.9.5",
     "yalc": "1.0.0-pre.53"
   },
   "overrides": {
