@@ -8,6 +8,7 @@ import * as graphqlUploadExpress from 'graphql-upload/graphqlUploadExpress.js';
 import { merge } from './core/common/helpers/config.helper';
 import { IServerOptions } from './core/common/interfaces/server-options.interface';
 import { MapAndValidatePipe } from './core/common/pipes/map-and-validate.pipe';
+import { ComplexityPlugin } from './core/common/plugins/complexity.plugin';
 import { ConfigService } from './core/common/services/config.service';
 import { EmailService } from './core/common/services/email.service';
 import { MailjetService } from './core/common/services/mailjet.service';
@@ -41,6 +42,13 @@ export class CoreModule implements NestModule {
    */
   static forRoot(AuthService: any, AuthModule: any, options: Partial<IServerOptions>): DynamicModule {
     // Process config
+    let cors = {};
+    if (options?.cookies) {
+      cors = {
+        credentials: true,
+        origin: true,
+      };
+    }
     const config: IServerOptions = merge(
       {
         env: 'develop',
@@ -52,7 +60,8 @@ export class CoreModule implements NestModule {
               Object.assign(
                 {
                   autoSchemaFile: 'schema.gql',
-                  context: ({ req }) => ({ req }),
+                  context: ({ req, res }) => ({ req, res }),
+                  cors,
                   installSubscriptionHandlers: true,
                   subscriptions: {
                     'subscriptions-transport-ws': {
@@ -135,6 +144,9 @@ export class CoreModule implements NestModule {
       EmailService,
       TemplateService,
       MailjetService,
+
+      // Plugins
+      ComplexityPlugin,
     ];
 
     // Return dynamic module
@@ -147,7 +159,7 @@ export class CoreModule implements NestModule {
         ),
       ],
       providers,
-      exports: [ConfigService, EmailService, TemplateService, MailjetService],
+      exports: [ConfigService, EmailService, TemplateService, MailjetService, ComplexityPlugin],
     };
   }
 }
