@@ -1,5 +1,8 @@
-import { Args, Info, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { Args, Context, Info, Mutation, Resolver } from '@nestjs/graphql';
+import { Response as ResponseType } from 'express';
 import { GraphQLResolveInfo } from 'graphql';
+import { ConfigService } from '../../../core/common/services/config.service';
+import { CoreAuthResolver } from '../../../core/modules/auth/core-auth.resolver';
 import { Auth } from './auth.model';
 import { AuthService } from './auth.service';
 import { AuthSignInInput } from './inputs/auth-sign-in.input';
@@ -9,21 +12,31 @@ import { AuthSignUpInput } from './inputs/auth-sign-up.input';
  * Authentication resolver for the sign in
  */
 @Resolver(() => Auth)
-export class AuthResolver {
+export class AuthResolver extends CoreAuthResolver {
   /**
    * Integrate services
    */
-  constructor(private readonly authService: AuthService) {}
+  constructor(
+    protected override readonly authService: AuthService,
+    protected override readonly configService: ConfigService
+  ) {
+    super(authService, configService);
+  }
 
   /**
    * SignIn for User
    */
-  @Query(() => Auth, { description: 'Sign in and get JWT token' })
-  async signIn(@Info() info: GraphQLResolveInfo, @Args('input') input: AuthSignInInput): Promise<Auth> {
-    return this.authService.signIn(input, {
+  @Mutation(() => Auth, { description: 'Sign in and get JWT token' })
+  override async signIn(
+    @Info() info: GraphQLResolveInfo,
+    @Context() ctx: { res: ResponseType },
+    @Args('input') input: AuthSignInInput
+  ): Promise<Auth> {
+    const result = await this.authService.signIn(input, {
       fieldSelection: { info, select: 'signIn' },
       inputType: AuthSignInInput,
     });
+    return this.processCookies(ctx, result);
   }
 
   /**
@@ -32,9 +45,14 @@ export class AuthResolver {
   @Mutation(() => Auth, {
     description: 'Sign up user and get JWT token',
   })
-  async signUp(@Info() info: GraphQLResolveInfo, @Args('input') input: AuthSignUpInput): Promise<Auth> {
-    return this.authService.signUp(input, {
+  override async signUp(
+    @Info() info: GraphQLResolveInfo,
+    @Context() ctx: { res: ResponseType },
+    @Args('input') input: AuthSignUpInput
+  ): Promise<Auth> {
+    const result = await this.authService.signUp(input, {
       fieldSelection: { info, select: 'signUp' },
     });
+    return this.processCookies(ctx, result);
   }
 }
