@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common';
+import { APP_INTERCEPTOR } from '@nestjs/core';
 import { ScheduleModule } from '@nestjs/schedule';
 import envConfig from '../config.env';
 import { CoreModule } from '../core.module';
+import { CheckSecurityInterceptor } from '../core/common/interceptors/check-security.interceptor';
 import { Any } from '../core/common/scalars/any.scalar';
 import { DateScalar } from '../core/common/scalars/date.scalar';
 import { JSON } from '../core/common/scalars/json.scalar';
@@ -34,7 +36,16 @@ import { ServerController } from './server.controller';
     FileModule,
   ],
 
-  providers: [Any, CronJobs, DateScalar, JSON],
+  providers: [
+    Any,
+    CronJobs,
+    DateScalar,
+    JSON,
+    {
+      provide: APP_INTERCEPTOR,
+      useClass: CheckSecurityInterceptor,
+    },
+  ],
 
   // Include REST controllers
   controllers: [ServerController],
