@@ -1,16 +1,20 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PubSub } from 'graphql-subscriptions';
 import { MongoClient, ObjectId } from 'mongodb';
-import { SortOrderEnum } from '../src';
+import { HttpExceptionLogFilter, SortOrderEnum, TestGraphQLType, TestHelper } from '../src';
 import envConfig from '../src/config.env';
 import { ComparisonOperatorEnum } from '../src/core/common/enums/comparison-operator.enum';
 import { RoleEnum } from '../src/core/common/enums/role.enum';
 import { User } from '../src/server/modules/user/user.model';
 import { UserService } from '../src/server/modules/user/user.service';
 import { ServerModule } from '../src/server/server.module';
-import { TestGraphQLType, TestHelper } from '../src/test/test.helper';
 
 describe('Project (e2e)', () => {
+  // To enable debugging, include these flags in the options of the request you want to debug
+  const log = true;
+  const logError = true;
+
+  // Testenvironment properties
   let app;
   let testHelper: TestHelper;
 
@@ -30,6 +34,10 @@ describe('Project (e2e)', () => {
    * Before all tests
    */
   beforeAll(async () => {
+    // Indicates that cookies are enabled
+    if (envConfig.cookies) {
+      console.error('NOTE: Cookie handling is enabled. The tests with tokens will fail!');
+    }
     try {
       const moduleFixture: TestingModule = await Test.createTestingModule({
         imports: [ServerModule],
@@ -42,6 +50,7 @@ describe('Project (e2e)', () => {
         ],
       }).compile();
       app = moduleFixture.createNestApplication();
+      app.useGlobalFilters(new HttpExceptionLogFilter());
       app.setBaseViewsDir(envConfig.templates.path);
       app.setViewEngine(envConfig.templates.engine);
       await app.init();
@@ -105,6 +114,7 @@ describe('Project (e2e)', () => {
     for (const user of users) {
       const res: any = await testHelper.graphQl({
         name: 'signIn',
+        type: TestGraphQLType.MUTATION,
         arguments: {
           input: {
             email: user.email,
