@@ -1,15 +1,20 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PubSub } from 'graphql-subscriptions';
 import { MongoClient, ObjectId } from 'mongodb';
-import { UserCreateInput } from '../src/server/modules/user/inputs/user-create.input';
-import { getPlain } from '../src/core/common/helpers/input.helper';
+import { HttpExceptionLogFilter, TestGraphQLType, TestHelper } from '../src';
 import envConfig from '../src/config.env';
+import { getPlain } from '../src/core/common/helpers/input.helper';
+import { UserCreateInput } from '../src/server/modules/user/inputs/user-create.input';
 import { User } from '../src/server/modules/user/user.model';
 import { UserService } from '../src/server/modules/user/user.service';
 import { ServerModule } from '../src/server/server.module';
-import { TestGraphQLType, TestHelper } from '../src/test/test.helper';
 
 describe('ServerModule (e2e)', () => {
+  // To enable debugging, include these flags in the options of the request you want to debug
+  const log = true;
+  const logError = true;
+
+  // Testenvironment properties
   const port = 3030;
   let app;
   let testHelper: TestHelper;
@@ -26,6 +31,7 @@ describe('ServerModule (e2e)', () => {
   let gEmail: string;
   let gPassword: string;
   let gToken: string;
+  let gRefreshToken: string;
 
   // ===================================================================================================================
   // Preparations
@@ -35,6 +41,10 @@ describe('ServerModule (e2e)', () => {
    * Before all tests
    */
   beforeAll(async () => {
+    // Indicates that cookies are enabled
+    if (envConfig.cookies) {
+      console.error('NOTE: Cookie handling is enabled. The tests with tokens will fail!');
+    }
     try {
       // Start server for testing
       const moduleFixture: TestingModule = await Test.createTestingModule({
@@ -48,6 +58,7 @@ describe('ServerModule (e2e)', () => {
         ],
       }).compile();
       app = moduleFixture.createNestApplication();
+      app.useGlobalFilters(new HttpExceptionLogFilter());
       app.setBaseViewsDir(envConfig.templates.path);
       app.setViewEngine(envConfig.templates.engine);
       await app.init();
@@ -109,7 +120,7 @@ describe('ServerModule (e2e)', () => {
           firstName: 'Everardo',
         },
       },
-      fields: [{ user: ['id', 'email', 'roles', 'createdBy'] }],
+      fields: ['token', 'refreshToken', { user: ['id', 'email', 'roles', 'createdBy'] }],
     });
     expect(res.user.email).toEqual(gEmail);
     expect(res.user.roles).toEqual([]);
@@ -198,17 +209,61 @@ describe('ServerModule (e2e)', () => {
   it('signIn', async () => {
     const res: any = await testHelper.graphQl({
       name: 'signIn',
+      type: TestGraphQLType.MUTATION,
       arguments: {
         input: {
           email: gEmail,
           password: gPassword,
         },
       },
-      fields: ['token', { user: ['id', 'email'] }],
+      fields: ['token', 'refreshToken', { user: ['id', 'email'] }],
     });
     expect(res.user.id).toEqual(gId);
     expect(res.user.email).toEqual(gEmail);
+    expect(res.token.length).toBeGreaterThan(0);
+    expect(res.refreshToken.length).toBeGreaterThan(0);
     gToken = res.token;
+    gRefreshToken = res.refreshToken;
+  });
+
+  /**
+   * Try to get refresh token with token
+   */
+  it('tryToGetRefreshTokenWithToken', async () => {
+    const res: any = await testHelper.graphQl(
+      {
+        name: 'refreshToken',
+        type: TestGraphQLType.MUTATION,
+        fields: ['token', 'refreshToken', { user: ['id', 'email'] }],
+      },
+      { token: gToken }
+    );
+    expect(res.errors.length).toBeGreaterThanOrEqual(1);
+    expect(res.errors[0].extensions.response.statusCode).toEqual(401);
+    expect(res.errors[0].message).toEqual('Unauthorized');
+    expect(res.data).toBe(null);
+  });
+
+  /**
+   * Get refresh token with refresh token
+   */
+  it('getRefreshTokenWithRefreshToken', async () => {
+    const res: any = await testHelper.graphQl(
+      {
+        name: 'refreshToken',
+        type: TestGraphQLType.MUTATION,
+        fields: ['token', 'refreshToken', { user: ['id', 'email'] }],
+      },
+      { token: gRefreshToken }
+    );
+    expect(res.user.id).toEqual(gId);
+    expect(res.user.email).toEqual(gEmail);
+    expect(res.token.length).toBeGreaterThan(0);
+    expect(res.refreshToken.length).toBeGreaterThan(0);
+    expect(res.token.length).not.toEqual(gToken);
+    expect(res.refreshToken.length).not.toEqual(gRefreshToken);
+    gToken = res.token;
+    gRefreshToken = res.refreshToken;
   });
 
   /**
